refactor: streamline utility functions and remove unused code in telomerehunter2_main.py and utils.py
