+++ src/telomerehunter2/telomerehunter2_main.py
@@ -21,11 +21,6 @@
 import argparse
 import multiprocessing
 import os
-import re
-import shutil
-import stat
-import sys
-import tempfile
 from concurrent.futures import ProcessPoolExecutor
 
 import numpy as np
@@ -39,20 +34,29 @@ from telomerehunter2 import (TVR_context, TVR_context_summary_tables,
                              normalize_TVR_counts, plot_functions,
                              repeat_frequency_intratelomeric,
                              sort_telomere_reads)
-
-source_directory = os.path.abspath(os.path.dirname(__file__))
-
+from telomerehunter2.utils import (
+    file_exists,
+    check_and_create_index,
+    delete_temp_dir,
+    check_and_prompt_filtering,
+    get_repeat_threshold_from_summary,
+    validate_args,
+    validate_plotting_options,
+    set_execution_flags,
+    combine_summary_files,
+    print_copyright_message,
+)
 
 def run_sample(
-    sample_name,
-    sample_bam,
-    outdir_sample,
-    sample_id,
-    read_length,
-    repeat_threshold_calc,
-    repeat_threshold_str,
-    filter_telomere_reads_flag,
-    args,
+        sample_name,
+        sample_bam,
+        outdir_sample,
+        sample_id,
+        read_length,
+        repeat_threshold_calc,
+        repeat_threshold_str,
+        filter_telomere_reads_flag,
+        args,
 ):
     if getattr(args, "singlecell_mode", False):
         print(f"[TH2] Running sample '{sample_name}' in single-cell mode.")
@@ -156,292 +160,6 @@ def run_sample(
                 context_after=args.bp_context,
                 repeat_threshold_set=args.repeat_threshold_set,
             )
-
-
-def print_copyright_message():
-    print("\n")
-    print(
-        "\tTelomereHunter2 Copyright 2024 Ferdinand Popp, Lina Sieverling, Philip Ginsbach, Chen Hong, Lars Feuerbach"
-    )
-    print("\tThis program comes with ABSOLUTELY NO WARRANTY.")
-    print("\tThis is free software, and you are welcome to redistribute it")
-    print(
-        "\tunder certain conditions. For details see the GNU General Public License v3.0"
-    )
-    print(
-        "\tin the license copy received with TelomereHunter2 or <http://www.gnu.org/licenses/>."
-    )
-    print("\n")
-    print("TelomereHunter2 1.0.0")
-    print("\n")
-
-
-def file_exists(parser, path):
-    if path is None:
-        return None
-    if not os.path.exists(path):
-        parser.error("The file %s does not exist!" % path)
-    return path
-
-
-def check_banding_file(banding_file, outdir=None):
-    """
-    Tests if the supplied tsv file has format like: chr1	0	2300000	p36.33	...
-    """
-    try:
-        # Read the file (assuming tab-separated or CSV)
-        try:
-            df = pd.read_csv(banding_file, sep="\t", header=None)
-        except:
-            df = pd.read_csv(banding_file, header=None)
-            raise ValueError(f"Banding file is not a tsv file.")
-
-        # Check if file has at least 4 columns
-        if df.shape[1] < 4:
-            raise ValueError(
-                f"File has fewer than 4 columns. Found {df.shape[1]} columns."
-            )
-
-        # Check column types for the first 4 columns
-        type_errors = []
-
-        # Check first column (should be string)
-        if not all(isinstance(x, str) for x in df[0].dropna()):
-            type_errors.append("Column 1 should contain strings (chromosome names)")
-
-        # Check columns 2 and 3 (should be integers)
-        for col_idx in [1, 2]:
-            # Convert to numeric if possible
-            df[col_idx] = pd.to_numeric(df[col_idx], errors="coerce")
-            if df[col_idx].isna().any():
-                type_errors.append(f"Column {col_idx + 1} should contain integers")
-
-        # Check column 4 (should be string)
-        if not all(isinstance(x, str) for x in df[3].dropna()):
-            type_errors.append("Column 4 should contain strings (band names)")
-
-        # If any type errors found, raise exception
-        if type_errors:
-            error_msg = "Invalid column types:\n" + "\n".join(
-                f"- {err}" for err in type_errors
-            )
-            raise ValueError(error_msg)
-
-        # Check if first column contains only unique values (chromosome-level data)
-        chromosome_level_only = df[0].nunique() == len(df)
-
-        if chromosome_level_only:
-            print("!!! Only chromosome-level data found, no band information. !!!")
-            print("The banding file should be discarded.")
-
-            # Generate band information if requested
-            subs_cytobands_df = generate_banding_file(df)
-            print(
-                "\nGenerated substitution cytoband information (5% from chromosome ends):"
-            )
-            print(subs_cytobands_df.head())
-            print(f"Total generated bands: {len(subs_cytobands_df)}")
-            generated_banding_file_path = (
-                f"{outdir}/generated_bandingfile_5percent_ends.txt"
-            )
-            subs_cytobands_df.to_csv(generated_banding_file_path, sep="\t", index=False)
-            print(
-                f"Saved cytoband information (5% from chromosome ends): {generated_banding_file_path}"
-            )
-
-            return generated_banding_file_path
-        else:
-            print("Band-level information found. File is valid for detailed analysis.")
-            return banding_file
-
-    except Exception as e:
-        print(f"ERROR: {str(e)}")
-        sys.exit(1)
-
-
-def generate_banding_file(df):
-    """
-    Generate band information for chromosomes by adding bands at 5% from the ends
-
-    Args:
-        df: DataFrame with chromosome-level data
-
-    Returns:
-        DataFrame with generated band information
-    """
-    bands_df = []
-
-    for _, row in df.iterrows():
-        chrom = row[0]
-        start = row[1]
-        end = row[2]
-
-        # Calculate positions at 5% from each end
-        band_width = end - start
-        p_terminal = start + int(0.05 * band_width)
-        q_terminal = end - int(0.05 * band_width)
-
-        # Create bands: p-terminal, middle, q-terminal
-        bands_df.append([chrom, start, p_terminal, f"{chrom}pter"])
-        bands_df.append([chrom, p_terminal, q_terminal, f"{chrom}pq"])
-        bands_df.append([chrom, q_terminal, end, f"{chrom}qter"])
-
-    return pd.DataFrame(bands_df)
-
-
-def get_bamfile_reference_genome(bam_file_path):
-    """
-    Extracts a specific string from the second line of the BAM file header.
-    """
-    try:
-        mode = "rb" if bam_file_path.endswith(".bam") else "rc"
-        with pysam.AlignmentFile(bam_file_path, mode=mode) as bam_file:
-            header = bam_file.header.to_dict()
-            reference_genome_string = header["SQ"][0]
-            print(f"Input file is based on reference genome: {reference_genome_string}")
-    except KeyError:
-        return "Error when extracting reference genome from bam file."
-
-
-def validate_args(args):
-    # check which bam files were specified
-    if not args.plot_mode and not any([args.tumor_bam, args.control_bam]):
-        raise ValueError(
-            "argument -ibt/--inputBamTumor or -ibc/--inputBamControl or -icc or -ict is required"
-        )
-
-    if not args.tumor_bam:
-        args.tumor_flag = False
-        print(
-            "Tumor BAM/CRAM file was not specified. Only running control BAM/CRAM file."
-        )
-    elif not args.control_bam:
-        args.control_flag = False
-        print(
-            "Control BAM/CRAM file was not specified. Only running tumor BAM/CRAM file."
-        )
-
-    # Check band file format
-    if args.banding_file:
-        if args.tumor_bam or args.control_bam:
-            get_bamfile_reference_genome(args.tumor_bam or args.control_bam)
-
-        # check banding file and if only chromosomes, create 5% bands
-        args.banding_file = check_banding_file(args.banding_file, args.outdir)
-    else:
-        print(
-            "No banding file supplied, so all banding steps will be skipped and specific plots omitted."
-        )
-
-    # check if repeats only contains ACGT and are between 4 and 9 bases
-    for repeat in args.repeats:
-        if len(repeat) < 4 or len(repeat) > 9:
-            raise ValueError(
-                "argument -r/--repeats should be between 4 and 9 bases long."
-            )
-        x = re.search(r"[^ACGT]", repeat)
-        if x is not None:
-            raise ValueError(
-                "argument -r/--repeats should only contain the letters ACGT."
-            )
-
-    for repeat in args.TVRs_for_context:
-        x = re.search(r"[^ACGT]", repeat)
-        if x is not None:
-            raise ValueError(
-                "argument -rc/--repeatsContext should only contain the letters ACGT."
-            )
-
-    # check if bp for sequencing context is divisible by 6 or base telomere sequence given
-    if args.bp_context % len(args.repeats[0]) != 0:
-        raise ValueError(
-            f"argument -bp/--bpContext must be a multiple of length of first telomere repeat sequence /"
-            f'{len(args.repeats[0])}. By default "TTAGGG" so multiple of 6 (e.g. 6, 12, 18, ...).'
-        )
-
-    if args.mapq_threshold < 0 or args.mapq_threshold > 40:
-        raise ValueError(
-            "argument -mqt/--mappingQualityThreshold must be an integer between 0 and 40."
-        )
-
-    if args.gc_lower < 0 or args.gc_lower > 100:
-        raise ValueError(
-            "argument -gc1/--lowerGC must be an integer between 0 and 100."
-        )
-
-    if args.gc_upper < 0 or args.gc_upper > 100:
-        raise ValueError(
-            "argument -gc2/--upperGC must be an integer between 0 and 100."
-        )
-
-    if args.gc_lower >= args.gc_upper:
-        raise ValueError(
-            "argument -gc1/--lowerGC must be less than argument -gc2/--upperGC."
-        )
-
-    if args.subsample:
-        if args.subsample < 0 or args.subsample > 1:
-            raise argparse.ArgumentTypeError(
-                f"Subsample value must be between 0 and 1. Got {args.subsample} instead."
-            )
-
-
-def validate_plotting_options(args):
-    if args.plotNone and (
-        args.plotChr
-        or args.plotFractions
-        or args.plotTelContent
-        or args.plotGC
-        or args.plotRepeatFreq
-        or args.plotTVR
-        or args.plotSingleton
-    ):
-        raise ValueError(
-            "argument -pno/--plotNone should not be specified when other plotting options are selected."
-        )
-
-    # can be str or list so change to list
-    if isinstance(args.plotFileFormat, str):
-        args.plotFileFormat = [args.plotFileFormat]
-    if args.plotFileFormat == "all":
-        args.plotFileFormat = ["pdf", "png", "svg"]
-
-    # if no plotting options are selected: plot all diagrams.
-    if (
-        not args.plotChr
-        and not args.plotFractions
-        and not args.plotTelContent
-        and not args.plotGC
-        and not args.plotRepeatFreq
-        and not args.plotTVR
-        and not args.plotSingleton
-        and not args.plotNone
-    ):
-        args.plotChr = True
-        args.plotFractions = True
-        args.plotTelContent = True
-        args.plotGC = True
-        args.plotRepeatFreq = True
-        args.plotTVR = True
-        args.plotSingleton = True
-    # no banding info remove plots
-    if not args.banding_file:
-        args.plotChr = False
-        args.plotFractions = False
-
-
-def set_execution_flags(args):
-    args.filter_telomere_reads_flag = not args.noFiltering
-    args.sort_telomere_reads_flag = args.estimate_telomere_content_flag = True
-    args.TVR_screen_flag = args.TVR_context_flag = True
-
-    if args.repeats[0] != "TTAGGG":
-        print(
-            "The first repeat to screen for is not canonical TTAGGG! Please double check."
-        )
-        args.TVR_screen_flag = args.TVR_context_flag = args.plotTVR = (
-            args.plotSingleton
-        ) = False
 
 
 def parse_command_line_arguments():
@@ -574,8 +292,8 @@ def parse_command_line_arguments():
         dest="repeats",
         type=str,
         help="Base sequences to inspect like TTAGGG and its TVRs. First sequence is /"
-        "used as base telomeric sequence. Default: ['TTAGGG', 'TGAGGG', 'TCAGGG', /"
-        "'TTCGGG', 'TTGGGG'].",
+             "used as base telomeric sequence. Default: ['TTAGGG', 'TGAGGG', 'TCAGGG', /"
+             "'TTCGGG', 'TTGGGG'].",
         default=["TTAGGG", "TGAGGG", "TCAGGG", "TTGGGG", "TTCGGG", "TTTGGG"],
     )
     repeats_context_group.add_argument(
@@ -745,148 +463,91 @@ def parse_command_line_arguments():
     return args
 
 
-def has_filtering_output(outdir, pid, sample_id):
-    sample_dir = os.path.join(outdir, f"{sample_id}_TelomerCnt_{pid}")
-    files = [
-        os.path.join(sample_dir, f"{pid}_filtered.bam"),
-        os.path.join(sample_dir, f"{pid}_filtered_name_sorted.bam"),
-        os.path.join(sample_dir, f"{pid}_readcount.tsv"),
-        os.path.join(sample_dir, f"{pid}_{sample_id}_gc_content.tsv"),
-    ]
-    return all(os.path.exists(file) for file in files)
-
-
-def check_and_prompt_filtering(
-    filter_telomere_reads_flag, tumor_flag, control_flag, outdir, pid
+def run_telomerehunter(
+        args,
+        control_bam,
+        filter_telomere_reads_control,
+        filter_telomere_reads_tumor,
+        read_lengths_control,
+        read_lengths_tumor,
+        repeat_thresholds_control,
+        repeat_thresholds_str_control,
+        repeat_thresholds_str_tumor,
+        repeat_thresholds_tumor,
+        tumor_bam,
 ):
-    tumor_output_missing = (
-        not has_filtering_output(outdir, pid, "tumor") if tumor_flag else False
-    )
-    control_output_missing = (
-        not has_filtering_output(outdir, pid, "control") if control_flag else False
-    )
+    # Determine maximum number of workers based on parallel execution setting
+    max_workers = 1 if not args.run_parallel else min(2, multiprocessing.cpu_count())
 
-    # Both missing
-    if tumor_output_missing and control_output_missing:
-        print("Running filtering step as no output files are present")
-        # Alert wrong flag use!
-        if filter_telomere_reads_flag:
-            print(
-                "!Denied skipping the filtering with --noFiltering, as no output files were present!"
+    with ProcessPoolExecutor(max_workers=max_workers) as executor:
+        submitted_futures = []
+
+        # run tumor sample for PID
+        if args.tumor_flag and (
+                filter_telomere_reads_tumor
+                or args.sort_telomere_reads_flag
+                or args.estimate_telomere_content_flag
+        ):
+            tumor_sample_id = "tumor"
+            tumor_output_dir = os.path.join(
+                args.outdir, f"{tumor_sample_id}_TelomerCnt_{args.pid}"
             )
-        filter_T, filter_C = True, True
-    # One missing due to only one run or only one present
-    elif control_output_missing or tumor_output_missing:
-        print("Running filtering step as not all output files are already present")
-        if filter_telomere_reads_flag:
-            print(
-                "!Denied skipping the filtering with --noFiltering, as not all output files were present!"
+            os.makedirs(tumor_output_dir, exist_ok=True)
+
+            # Submit to run_sample main function
+            tumor_future = executor.submit(
+                run_sample,
+                sample_name="Tumor Sample",
+                sample_bam=tumor_bam,
+                outdir_sample=tumor_output_dir,
+                sample_id=tumor_sample_id,
+                read_length=read_lengths_tumor,
+                repeat_threshold_calc=repeat_thresholds_tumor,
+                repeat_threshold_str=repeat_thresholds_str_tumor,
+                filter_telomere_reads_flag=filter_telomere_reads_tumor,
+                args=args,
             )
-        filter_T, filter_C = tumor_output_missing, control_output_missing
-    # Both present
-    else:
-        if filter_telomere_reads_flag:
-            print(
-                "Filtering samples, but output files are already present. Consider using the --noFiltering flag"
+
+            submitted_futures.append(tumor_future)
+
+            if not args.run_parallel:
+                tumor_future.result()
+
+        # run control samples for PID
+        if args.control_flag and (
+                filter_telomere_reads_control
+                or args.sort_telomere_reads_flag
+                or args.estimate_telomere_content_flag
+        ):
+            control_sample_id = "control"
+            control_output_dir = os.path.join(
+                args.outdir, f"{control_sample_id}_TelomerCnt_{args.pid}"
             )
-            filter_T, filter_C = tumor_flag, control_flag
-        else:
-            print("Skipped filtering samples with --noFiltering flag")
-            filter_T, filter_C = False, False
+            os.makedirs(control_output_dir, exist_ok=True)
 
-    return filter_T, filter_C
+            # Submit to run_sample main function
+            control_future = executor.submit(
+                run_sample,
+                sample_name="Control Sample",
+                sample_bam=control_bam,
+                outdir_sample=control_output_dir,
+                sample_id=control_sample_id,
+                read_length=read_lengths_control,
+                repeat_threshold_calc=repeat_thresholds_control,
+                repeat_threshold_str=repeat_thresholds_str_control,
+                filter_telomere_reads_flag=filter_telomere_reads_control,
+                args=args,
+            )
 
+            submitted_futures.append(control_future)
 
-def get_repeat_threshold_from_summary(args):
-    summary_file = os.path.join(args.outdir, f"{args.pid}_summary.tsv")
-    try:
-        # Read the summary file
-        summary = pd.read_csv(summary_file, sep="\t")
+            if not args.run_parallel:
+                control_future.result()
 
-        # Ensure the column exists
-        if "repeat_threshold_used" in summary.columns:
-            # Get unique values in the column
-            unique_thresholds = summary["repeat_threshold_used"].dropna().unique()
-
-            # Check the number of unique values
-            if len(unique_thresholds) == 1:
-                return str(int(unique_thresholds[0]))
-            elif len(unique_thresholds) > 1:
-                return "n"
-        else:
-            print(f"Error: Column 'repeat_threshold_used' not found in {summary_file}.")
-            return "n"
-
-    except Exception as e:
-        print(f"Error reading file {summary_file}: {e}")
-        return "n"
-
-
-def combine_summary_files(outdir, pid, tumor_flag, control_flag):
-    # Define the path for the combined summary file
-    summary_path = os.path.join(outdir, f"{pid}_summary.tsv")
-
-    tumor_summary_path = os.path.join(
-        outdir, f"tumor_TelomerCnt_{pid}", f"{pid}_tumor_summary.tsv"
-    )
-    control_summary_path = os.path.join(
-        outdir, f"control_TelomerCnt_{pid}", f"{pid}_control_summary.tsv"
-    )
-
-    # Check if both tumor and control flags are True
-    if tumor_flag and control_flag:
-        # Copy the tumor summary file to the combined summary file
-        shutil.copyfile(tumor_summary_path, summary_path)
-
-        # Append the last line of the control summary file to the combined summary file
-        with open(control_summary_path, "r") as control_file, open(
-            summary_path, "a"
-        ) as combined_file:
-            last_line = control_file.readlines()[-1]
-            combined_file.write(last_line)
-
-    # Check if only the tumor flag is True
-    elif tumor_flag:
-        # Copy the tumor summary file to the combined summary file
-        shutil.copyfile(tumor_summary_path, summary_path)
-    # Check if only the control flag is True
-    elif control_flag:
-        # Copy the control summary file to the combined summary file
-        shutil.copyfile(control_summary_path, summary_path)
-
-
-def check_and_create_index(bam_path):
-    # make path absolute
-    bam_path = os.path.abspath(bam_path)
-
-    # Check if index file exists
-    index_path = bam_path + ".bai"
-    if os.path.exists(index_path):
-        return_bam = bam_path  # Index file already exists, no need to create
-    else:
-        # If the directory is writeable
-        if os.access(os.path.dirname(bam_path), os.W_OK):
-            pysam.index(bam_path)
-            return_bam = bam_path
-        else:
-            # Not writeable needs temp dir
-            temp_dir = os.path.abspath(tempfile.mkdtemp("tmp_TH2"))
-
-            # Set read and write permissions for the temporary directory
-            os.chmod(temp_dir, stat.S_IRUSR | stat.S_IWUSR | stat.S_IXUSR)
-
-            temp_bam_path = os.path.join(temp_dir, os.path.basename(bam_path))
-
-            # symbolic link bam file in temp folder
-            try:
-                os.symlink(bam_path, temp_bam_path)
-            except OSError as e:
-                print(f"Symbolic link creation failed: {e}")
-
-            pysam.index(temp_bam_path)
-            return_bam = temp_bam_path
-
-    return return_bam
+        # Wait for all futures to complete (only needed in parallel mode)
+        if args.run_parallel:
+            for future in submitted_futures:
+                future.result()
 
 
 def get_read_lengths_and_repeat_thresholds(args, control_bam, tumor_bam):
@@ -973,109 +634,6 @@ def get_read_lengths_and_repeat_thresholds(args, control_bam, tumor_bam):
         repeat_thresholds_str_tumor,
         repeat_thresholds_tumor,
     )
-
-
-def run_telomerehunter(
-    args,
-    control_bam,
-    filter_telomere_reads_control,
-    filter_telomere_reads_tumor,
-    read_lengths_control,
-    read_lengths_tumor,
-    repeat_thresholds_control,
-    repeat_thresholds_str_control,
-    repeat_thresholds_str_tumor,
-    repeat_thresholds_tumor,
-    tumor_bam,
-):
-    # Determine maximum number of workers based on parallel execution setting
-    max_workers = 1 if not args.run_parallel else min(2, multiprocessing.cpu_count())
-
-    with ProcessPoolExecutor(max_workers=max_workers) as executor:
-        submitted_futures = []
-
-        # run tumor sample for PID
-        if args.tumor_flag and (
-            filter_telomere_reads_tumor
-            or args.sort_telomere_reads_flag
-            or args.estimate_telomere_content_flag
-        ):
-            tumor_sample_id = "tumor"
-            tumor_output_dir = os.path.join(
-                args.outdir, f"{tumor_sample_id}_TelomerCnt_{args.pid}"
-            )
-            os.makedirs(tumor_output_dir, exist_ok=True)
-
-            # Submit to run_sample main function
-            tumor_future = executor.submit(
-                run_sample,
-                sample_name="Tumor Sample",
-                sample_bam=tumor_bam,
-                outdir_sample=tumor_output_dir,
-                sample_id=tumor_sample_id,
-                read_length=read_lengths_tumor,
-                repeat_threshold_calc=repeat_thresholds_tumor,
-                repeat_threshold_str=repeat_thresholds_str_tumor,
-                filter_telomere_reads_flag=filter_telomere_reads_tumor,
-                args=args,
-            )
-
-            submitted_futures.append(tumor_future)
-
-            if not args.run_parallel:
-                tumor_future.result()
-
-        # run control samples for PID
-        if args.control_flag and (
-            filter_telomere_reads_control
-            or args.sort_telomere_reads_flag
-            or args.estimate_telomere_content_flag
-        ):
-            control_sample_id = "control"
-            control_output_dir = os.path.join(
-                args.outdir, f"{control_sample_id}_TelomerCnt_{args.pid}"
-            )
-            os.makedirs(control_output_dir, exist_ok=True)
-
-            # Submit to run_sample main function
-            control_future = executor.submit(
-                run_sample,
-                sample_name="Control Sample",
-                sample_bam=control_bam,
-                outdir_sample=control_output_dir,
-                sample_id=control_sample_id,
-                read_length=read_lengths_control,
-                repeat_threshold_calc=repeat_thresholds_control,
-                repeat_threshold_str=repeat_thresholds_str_control,
-                filter_telomere_reads_flag=filter_telomere_reads_control,
-                args=args,
-            )
-
-            submitted_futures.append(control_future)
-
-            if not args.run_parallel:
-                control_future.result()
-
-        # Wait for all futures to complete (only needed in parallel mode)
-        if args.run_parallel:
-            for future in submitted_futures:
-                future.result()
-
-
-def delete_temp_dir(tumor_bam, control_bam, temp_dir):
-    temp_exist = [
-        os.path.dirname(path)
-        for path in [tumor_bam, control_bam]
-        if path is not None and temp_dir in path
-    ]
-
-    if temp_exist:
-        try:
-            # Attempt to remove the folder and its contents (same for both)
-            shutil.rmtree(temp_exist[0])
-            print(f"Folder '{temp_exist}' successfully deleted.")
-        except Exception as e:
-            print(f"Error deleting folder '{temp_exist}': {e}")
 
 
 def summary_log2(main_path, pid):
